refactor: replace TooltipAnimationWrapper with AnimationFrame for improved tooltip animation handling
